refactor: clean-up types, move presets service to the correct type folder
