+++ packages/vuetify/src/presets/default/index.ts
@@ -9,7 +9,7 @@ import { en } from '../../locale'
 // import mdi from '../../services/icons/presets/mdi'
 
 // Types
-import { VuetifyPreset } from 'vuetify/types/presets'
+import { VuetifyPreset } from 'vuetify/types/services/presets'
 
 export const Preset: VuetifyPreset = {
   breakpoint: {
