+++ packages/vuetify/types/services/index.d.ts
@@ -1,7 +1,7 @@
 // Types
 import Vue from 'vue'
 import Framework from '../'
-import { VuetifyPreset } from 'vuetify/types/presets'
+import { VuetifyPreset } from 'vuetify/types/services/presets'
 
 export interface VuetifyServiceContract {
   framework: Record<string, VuetifyServiceContract>
